made package object utils
