+++ src/main/scala/it/unibo/pps1920/motoscala/ecs/util/Alignment.scala
@@ -1,0 +1,3 @@
+package it.unibo.pps1920.motoscala.ecs.util
+
+sealed case class Alignment(var horizontalSide: Side, var verticalSide: Side)
