+++ src/test/scala/it/unibo/pps1920/motoscala/ecs/systems/CollisionSystemTest.scala
@@ -16,6 +16,7 @@ import org.scalatest.BeforeAndAfterAll
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 import org.scalatestplus.junit.JUnitRunner
+import it.unibo.pps1920.motoscala.ecs.systems.Collision.CollisionsSystem
 
 @RunWith(classOf[JUnitRunner])
 class CollisionSystemTest extends AnyWordSpec with Matchers with BeforeAndAfterAll {
